ai-services/llm-gateway/requirements.txt
fastapi==0.104.1
litellm==1.35.8
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
loguru==0.7.2
